requirements_postgresql.txt
"""
Requirements for PostgreSQL-based hybrid QA system
"""

# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# PostgreSQL database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# AI/ML components
langchain==0.0.339
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.35.2
torch==2.1.1

# File processing
pypdf2==3.0.1
python-docx==1.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.24.4
pandas==2.1.4

# Development/Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: For advanced text processing
spacy==3.7.2
nltk==3.8.1